+++ minion/include/abc.h
@@ -39,7 +39,7 @@ public:
      * @brief Constructor for ABC.
      * @param func The objective function to minimize.
      * @param bounds The bounds for the variables.
-     * @param x0 The initial solution.
+     * @param x0 The initial solution. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. These guesses will be used for population initialization.
      * @param data Additional data for the objective function.
      * @param callback Callback function for intermediate results.
      * @param tol The tolerance for stopping criteria.