+++ minion/include/jso.h
@@ -33,7 +33,7 @@ class jSO : public Differential_Evolution {
          * @param func The objective function to minimize.
          * @param bounds The bounds for the variables.
          * @param options A map of configuration options.
-         * @param x0 The initial guesses for the solution.
+         * @param x0 The initial guesses for the solution. Note that Minion assumes multiple initial guesses, thus, x0 is an std::vector<std::vector<double>> object. These guesses will be used for population initialization.
          * @param data Additional data for the objective function.
          * @param callback Callback function for intermediate results.
          * @param tol The tolerance for stopping criteria.