+++ docs/source/changelog.rst
@@ -5,6 +5,20 @@ Changelog
    :local:
    :depth: 2
 
+Version 0.2.3 - 2025-08-21
+--------------------------
+
+**Changed**
+
+- Classical algorithms such as L-BFGS pick the best guess from the given guesses during initialization. 
+- Stability improvements in L-BFGS and L-BFGS-B
+
+Version 0.2.2 - 2025-07-31
+--------------------------
+
+**Changed**
+
+- Initial guess x0 is now a list of guesses. Thus, it is possible now to have more than one initial guesses.
 
 Version 0.2.1 - 2025-03-01
 --------------------------